Anpassung des Kontigent auf Long wurde in der Buyactivity vergessen

+++ app/src/main/java/com/example/david/cheapfood/Offer.java
@@ -34,7 +34,7 @@ public class Offer {
         return contingent;
     }
 
-    public void setContigent(int contigent) {
+    public void setContigent(long contigent) {
         this.contingent = contigent;
     }
